+++ lib/domain/entities/product.dart
@@ -1,5 +1,3 @@
-import 'package:product_management/data/models/product_model.dart';
-
 class Product {
   final int id;
   final String name;
