+++ lib/domain/usecases/usecases.dart
@@ -1,3 +1,5 @@
 export 'login_usecase.dart';
 export 'fetch_products.dart';
 export 'fetch_product_detail.dart';
+export 'update_product.dart';
+export 'create_product.dart';
