+++ lib/presentation/views/product_detail_page.dart
@@ -9,8 +9,7 @@ import 'package:product_management/presentation/views/product_form_page.dart';
 
 class ProductDetailPage extends StatefulWidget {
   final int productId;
-  const ProductDetailPage({Key? key, required this.productId})
-    : super(key: key);
+  const ProductDetailPage({super.key, required this.productId});
 
   @override
   State<ProductDetailPage> createState() => _ProductDetailPageState();
