+++ lib/presentation/views/views.dart
@@ -1,1 +1,3 @@
 export 'login_page.dart';
+export 'product_list_page.dart';
+export 'product_detail_page.dart';
